% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheit convertor}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{a numeric input of temperature in Fahrenheit to be convert to Celsius.}
}
\value{
A numeric that gives the temp in Celsius.
}
\description{
This function convert Fahrenheit to Celsius.
}
\keyword{temperature}
