% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pow.R
\name{my_pow}
\alias{my_pow}
\title{Power Function}
\usage{
my_pow(x, power = 2)
}
\arguments{
\item{x}{A numeric input to be raised by the power of the \code{power}.}

\item{power}{a numeric input for the power to raise \code{x},
the default is \code{2}.}
}
\value{
A numeric that gives the result of \code{x} raised to the power of \code{power}.
}
\description{
This function computes the power of the input.
}
